se agrego login, mazo, funcionalidades
