+++ proyectSrc/react-front/src/components/componentsLiq/PanelControl.js
